refactoring observation
