+++ ngmix/gmix/gmix.py
@@ -31,7 +31,7 @@ from ..gmix.gmix_nb import (
 )
 
 from .render_nb import render
-from ..pixels import make_coords
+from ..observation.pixels import make_coords
 
 
 def make_gmix_model(pars, model, **kw):
